rename /middleware/tools <- /res
